pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-complexity-analyzer"
version = "0.1.0"
description = "Agentic Complexity Analyzer V1"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.41",
  "psycopg[binary]>=3.2.9",
  "alembic>=1.16.0",
  "celery[redis]>=5.5.3",
  "redis>=6.4.0",
  "python-dotenv>=1.1.1",
  "pyyaml>=6.0.2",
  "networkx>=3.5",
  "radon>=6.0.1",
  "libcst>=1.8.5",
  "numpy>=2.3.0",
  "pandas>=2.3.0",
  "pyarrow>=20.0.0",
  "markdown>=3.8.2",
  "jinja2>=3.1.6",
  "python-multipart>=0.0.20",
  "httpx>=0.28.1",
  "psutil>=7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-cov>=6.2.1",
  "ruff>=0.12.8",
  "mypy>=1.17.1",
]

[project.scripts]
api-server = "apps.api.main:run"
worker-runner = "workers.sandbox_runner.main:run_worker"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "services*", "workers*", "analyzers*", "common*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
